+++ backend/app/domains/token/schemas.py
@@ -34,3 +34,11 @@ class UserTokenResponse(BaseModel):
 
     class Config:
         orm_mode = True
+
+class TokenInfo(BaseModel):
+    total_tokens: int
+    used_tokens: int
+    last_charged_at: Optional[datetime]
+
+    class Config:
+        orm_mode = True
